requirements.txt
python-dotenv
sentence-transformers
torch
pymilvus
psycopg[binary]
redis
#langchain-google-genai
#langchain
#langchain-community==0.2.6
#langgraph
sigmatools>=0.15.0
